% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3heatmap.R
\name{d3heatmap}
\alias{d3heatmap}
\title{d3heatmap}
\usage{
d3heatmap(x, Rowv = TRUE, Colv = if (symm) "Rowv" else TRUE,
  distfun = dist, hclustfun = hclust, dendrogram = c("both", "row",
  "column", "none"), reorderfun = function(d, w) reorder(d, w), k_row, k_col,
  symm = FALSE, revC, scale = c("none", "row", "column"), na.rm = TRUE,
  labRow = rownames(x), labCol = colnames(x), cexRow, cexCol, digits = 3L,
  cellnote, cellnote_scale = FALSE, theme = NULL, colors = "RdYlBu",
  width = NULL, height = NULL, xaxis_height = 80, yaxis_width = 120,
  xaxis_font_size = NULL, yaxis_font_size = NULL, brush_color = "#0000FF",
  show_grid = TRUE, anim_duration = 500, ...)
}
\description{
d3heatmap
}
\examples{
d3heatmap(mtcars, scale = "column", colors = "Blues")
}
