% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT.R
\name{datatable}
\alias{datatable}
\title{DataTable}
\usage{
datatable(data, options = list(), class = "display",
  callback = JS("return table;"), rownames, colnames, container,
  caption = NULL, filter = c("none", "bottom", "top"), escape = TRUE,
  style = "default", width = NULL, height = NULL, elementId = NULL,
  fillContainer = getOption("DT.fillContainer", NULL),
  autoHideNavigation = getOption("DT.autoHideNavigation", NULL),
  selection = c("multiple", "single", "none"), extensions = list(),
  plugins = NULL)
}
\description{
DataTable
}
\examples{
datatable(iris)

datatable(head(iris), class = 'cell-border stripe')

iris2 = iris[c(1:10, 51:60, 101:110), ]
datatable(iris2, filter = 'top', options = list(
  pageLength = 5, autoWidth = TRUE
))
}
