% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{ggplotly}
\alias{ggplotly}
\title{ggplotly}
\usage{
ggplotly(p = ggplot2::last_plot(), width = NULL, height = NULL,
  tooltip = "all", dynamicTicks = FALSE, layerData = 1,
  originalData = TRUE, source = "A", ...)
}
\description{
ggplotly
}
\examples{
library(ggplot2)

data(canada.cities, package = "maps")
viz <- ggplot(canada.cities, aes(long, lat)) +
  borders(regions = "canada") +
  coord_equal() +
  geom_point(aes(text = name, size = pop), colour = "red", alpha = 1/2)
viz

p <- ggplotly(viz, tooltip = c("text", "size"))
p$elementId <- "ggplotly"
p$x$cur_data <- p$elementId
p
}
