% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly.R
\name{plot_ly}
\alias{plot_ly}
\title{plot_ly}
\usage{
plot_ly(data = data.frame(), ..., type = NULL, color, colors = NULL,
  alpha = 1, symbol, symbols = NULL, size, sizes = c(10, 100), linetype,
  linetypes = NULL, split, frame, width = NULL, height = NULL,
  source = "A")
}
\description{
plot_ly
}
\examples{
library(ggplot2)
p <- plot_ly(economics, x = ~date, y = ~pop)
p$elementId <- "plotly"
p
}
